update: double to big decimal

+++ src/main/java/com/mobiquity/entity/Item.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 @ToString
 public class Item {
     private final int index;
-    private final double weight;
-    private final double price;
+    private final BigDecimal weight;
+    private final BigDecimal price;
 }
